Reogranize package res
